Fixed the "groups" property of the SharedFolder class
Rather than being an array of strings the "groups" property is an array of JSON objects in a fomat very similar to the Membership/User format. This introduces those objects. Without this change any result that includes a SharedFolder object with a populated "groups" property will fail with a JSON parsing error.